Set software to remove even it is not installed. Ticket#2019012967001319 — MRR OSS4 - UltraVNC deinstalliert nicht und verursacht flackernde Bildschirme

+++ oss-dao/src/main/java/de/openschoolserver/dao/controller/SoftwareController.java
@@ -1419,12 +1419,14 @@ public class SoftwareController extends Controller {
 
 			//Remove first the softwares.
 			if( softwaresToRemove.containsKey(device.getName()) ) {
+				List<Software> removed     = new ArrayList<Software>();
 				for( Software software : softwaresToRemove.get(device.getName()) ) {
-					if( this.isSoftwareInstalledOnDevice(device, software) ){
+					if(!removed.contains(software)) {
 						this.setSoftwareStatusOnDevice(device, software, "", "DS");
 						deviceRemove.add("       - " + software.getName());
+						this.deleteSoftwareLicenseFromDevice(software,device);
+						removed.add(software);
 					}
-					this.deleteSoftwareLicenseFromDevice(software,device);
 				}
 			}
 			if( softwaresToInstall.containsKey(device.getName()) ) {
